Modified script
- Made the 'ExportPath' parameter optional.
- Improved performance by switching from a traditional array to a
'List<t>' object.
- Data will be returned to the console.
	- Doing this will allow for the ability to export the data how
	the user wants it to.
- Updated the help section to reflect the new changes.

+++ SharePointMigrationTool-Scripts/Invoke-UserHomeMigrationBuilder.ps1
@@ -10,19 +10,32 @@
 .PARAMETER SPOSubFolder
     The sub-folder to put the migrated data into during the migration job. Useful for preventing conflicts that may already exist.
 .PARAMETER ExportPath
-    The file path to export the CSV file to.
+    The file path to export a CSV file of the data to.
 .EXAMPLE
     PS C:\> .\Invoke-UserHomeMigrationBuilder.ps1 -UserName @("jdoe1", "jwinger", "pryan") -TenantName "contoso.com" -SPOSubFolder "UserHome Migration" -ExportPath ".\UserHomeDir-MigrationJob.csv"
+
+    Export the data to a CSV file to use in the SharePoint Migration Tool GUI interface.
+
+.EXAMPLE
+    PS C:\> .\Invoke-UserHomeMigrationBuilder.ps1 -UserName @("jdoe1", "jwinger", "pryan") -TenantName "contoso.com" -SPOSubFolder "UserHome Migration"
+
+    Output configuration only to the console.
+.EXAMPLE
+    PS C:\> .\Invoke-UserHomeMigrationBuilder.ps1 -UserName @("jdoe1", "jwinger", "pryan") -TenantName "contoso.com" -SPOSubFolder "UserHome Migration" | ConvertTo-Json | Out-File -FilePath ".\UserHomeDir-MigrationJob.json"
+
+    Output the configuration to a JSON file.
 #>
 [CmdletBinding()]
 param(
     [Parameter(Position = 0, Mandatory)]
     [string[]]$UserName,
+    [Parameter(Position = 1)]
+    [string]$AdServer,
     [Parameter(Position = 1, Mandatory)]
     [string]$TenantName,
     [Parameter(Position = 2)]
     [string]$SPOSubFolder,
-    [Parameter(Position = 2, Mandatory)]
+    [Parameter(Position = 3)]
     [string]$ExportPath
 )
 
@@ -36,26 +49,38 @@ begin {
         [string]$TargetSubFolder
     }
 
+    #Putting the common parameters being used in the `Get-ADUser` cmdlet.
+    $AdUserSplat = @{
+        "Properties" = @("HomeDirectory");
+    }
+
+    if ($null -ne $AdServer) {
+        #If the 'AdServer' parameter was provided, add it to the $AdUserSplat hashtable.
+        $AdUserSplat.Add("Server", $AdServer)
+    }
+
     $DomainSharepointBase = $TenantName.Replace(".", "") #Remove the '.' character for the sub-domain name to the Sharepoint Online domain.
     $SharepointPersonalDomain = $TenantName.Replace(".", "_") #Replace the '.' character with '_' to map to a user's personal Sharepoint Online library (OneDrive).
 }
 
 process {
-    $UserData = @()
+    $UserData = [System.Collections.Generic.List[UserAccountObj]]::new()
 
     #Loop through each UserName provided.
     foreach ($User in $UserName) {
         Write-Verbose "Getting user data for '$($User)'."
-        $UserAccount = Get-ADUser -Identity $User -Properties "HomeDirectory" #Collect the user's AD object and ensure that the 'HomeDirectory' property is returned.
+        $UserAccount = Get-ADUser -Identity $User @AdUserSplat #Collect the user's AD object and ensure that the 'HomeDirectory' property is returned.
 
         if ($UserAccount.HomeDirectory) {
             #If the 'HomeDirectory' property does exist on the AD object, build an object and add it to the UserData array
-            $UserData += [UserAccountObj]@{
-                "Source"          = $UserAccount.HomeDirectory;
-                "TargetWeb"       = "https://$($DomainSharepointBase)-my.sharepoint.com/personal/$($UserAccount.Name)_$($SharepointPersonalDomain)/";
-                "TargetDocLib"    = "Documents";
-                "TargetSubFolder" = $SPOSubFolder;
-            }
+            $UserData.Add(
+                [UserAccountObj]@{
+                    "Source"          = $UserAccount.HomeDirectory;
+                    "TargetWeb"       = "https://$($DomainSharepointBase)-my.sharepoint.com/personal/$($UserAccount.Name)_$($SharepointPersonalDomain)/";
+                    "TargetDocLib"    = "Documents";
+                    "TargetSubFolder" = $SPOSubFolder;
+                }
+            )
         }
         else {
             #If the 'HomeDirectory' property does not exist on the AD object, return a warning message to the console. 
@@ -63,13 +88,18 @@ process {
         }
     }
 
-    $TmpFile = New-TemporaryFile #Create a temporary file to export the data to
-    Write-Verbose "Exporting the data to '$($ExportPath)'."
-    $UserData | Export-Csv -Path $TmpFile.FullName -NoTypeInformation #Export the data as a CSV file to the temporary file
-    $ExportedData = Get-Content -Path $TmpFile.FullName #Get the file content of the temporary file
-    $null = Remove-Item -Path $TmpFile -Force #Remove the temporary file from the system
+    if (([String]::IsNullOrEmpty($ExportPath)) -eq $false) {
+        $TmpFile = New-TemporaryFile #Create a temporary file to export the data to
+        Write-Verbose "Exporting the data to '$($ExportPath)'."
+        $UserData | Export-Csv -Path $TmpFile.FullName -NoTypeInformation #Export the data as a CSV file to the temporary file
+        $ExportedData = Get-Content -Path $TmpFile.FullName #Get the file content of the temporary file
+        $null = Remove-Item -Path $TmpFile -Force #Remove the temporary file from the system
 
-    #For some odd reason, the Sharepoint Migration Tool does not like having the header line in the CSV file. When we export the final file, it just skips the first line and saves it to the path specified.
-    $ExportedData | Select-Object -Skip 1 | Out-File -FilePath $ExportPath
+        #For some odd reason, the Sharepoint Migration Tool does not like having the header line in the CSV file. When we export the final file, it just skips the first line and saves it to the path specified.
+        $ExportedData | Select-Object -Skip 1 | Out-File -FilePath $ExportPath
+    }
+}
 
+end {
+    return $UserData
 }
